moved currency selector dropdown into a reusable component since most of the code was shared between the mobile and desktop version
